user_interface/app/requirements.txt
django
djangorestframework
pillow
numpy
tensorflow
transformers
whitenoise
gunicorn